requirements.txt
torchvision>=0.5.0
pretrainedmodels==0.7.4
efficientnet-pytorch==0.7.1
timm==0.9.2
tqdm==4.65.0
Pillow==9.3.0
opencv-python==4.8.0.74
clearml==1.11.1
albumentations==1.3.1
pytorch_lightning==2.0.5
matplotlib==3.7.2
tensorboard==2.13.0
onnx
